Added LogoutSpot Module

+++ BleachHack-Fabric-1.16/src/main/java/bleach/hack/gui/EntityMenuScreen.java
@@ -56,8 +56,8 @@ public class EntityMenuScreen extends Screen {
 	}
 
 	private void cursorMode(int mode) {
-		double x = (double)(this.client.getWindow().getWidth() / 2);
-		double y = (double)(this.client.getWindow().getHeight() / 2);
+		double x = this.client.getWindow().getWidth() / 2;
+		double y = this.client.getWindow().getHeight() / 2;
 
 		KeyBinding.unpressAll();
 		InputUtil.setCursorParameters(this.client.getWindow().getHandle(), GLFW.GLFW_CURSOR_HIDDEN, x, y);
